pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "fast-yaml"
version = "0.1.0"
description = "A fast YAML parser for Python, powered by Rust"
readme = "README.md"
license = { text = "MIT OR Apache-2.0" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Rust",
    "Topic :: Text Processing :: Markup",
]
keywords = ["yaml", "parser", "serialization", "rust"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/fast-yaml"
Repository = "https://github.com/yourusername/fast-yaml"
Documentation = "https://github.com/yourusername/fast-yaml#readme"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "fast_yaml._core"

[tool.pytest.ini_options]
testpaths = ["tests"]
